@@ -5,7 +5,7 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build -prod --output-path ui-src/META-INF/resources",
+    "build": "ng build --prod --output-path ui-src/META-INF/resources",
     "test": "ng test --watch false",
     "test-watch": "ng test",
     "lint": "ng lint",
@@ -15,51 +15,53 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/core": "^0.6.5",
-    "@angular/animations": "^5.2.11",
-    "@angular/cdk": "^5.2.5",
-    "@angular/common": "^5.2.11",
-    "@angular/compiler": "^5.2.11",
-    "@angular/core": "^5.2.11",
-    "@angular/forms": "^5.2.11",
-    "@angular/http": "^5.2.11",
-    "@angular/material": "^5.2.5",
-    "@angular/platform-browser": "^5.2.11",
-    "@angular/platform-browser-dynamic": "^5.2.11",
-    "@angular/router": "^5.2.11",
-    "core-js": "^2.5.7",
-    "file-saver": "^1.3.8",
+    "@angular-devkit/core": "7.1.4",
+    "@angular/animations": "7.1.4",
+    "@angular/cdk": "^7.2.0",
+    "@angular/common": "7.1.4",
+    "@angular/compiler": "7.1.4",
+    "@angular/core": "7.1.4",
+    "@angular/forms": "7.1.4",
+    "@angular/http": "7.1.4",
+    "@angular/material": "^7.2.0",
+    "@angular/platform-browser": "7.1.4",
+    "@angular/platform-browser-dynamic": "7.1.4",
+    "@angular/router": "7.1.4",
+    "core-js": "^2.6.1",
+    "file-saver": "^2.0.0",
     "font-awesome": "^4.7.0",
-    "ngx-gallery": "^5.3.2",
-    "ngx-page-scroll": "^4.0.2",
-    "npm": "^6.1.0",
-    "primeng": "^5.2.7",
-    "rxjs": "^5.5.11",
-    "snyk": "^1.108.2",
+    "ngx-gallery": "^5.9.0",
+    "ngx-page-scroll": "^5.0.1",
+    "npm": "^6.5.0",
+    "primeng": "^7.0.3",
+    "rxjs": "^6.3.3",
+    "snyk": "^1.118.2",
+    "tslib": "^1.9.3",
     "web-animations-js": "^2.3.1",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.5",
-    "@angular/compiler-cli": "^5.2.11",
-    "@angular/language-service": "^5.2.11",
+    "@angular-devkit/build-angular": "~0.11.0",
+    "@angular/cli": "7.1.4",
+    "@angular/compiler-cli": "7.1.4",
+    "@angular/language-service": "7.1.4",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.111",
     "angular-ide": "^0.9.51",
-    "codelyzer": "^4.3.0",
-    "jasmine-core": "~3.1.0",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~2.0.4",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
+    "codelyzer": "^4.4.0",
+    "jasmine-core": "~3.3.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~3.1.4",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-cli": "~2.0.0",
     "karma-coverage-istanbul-reporter": "^1.4.3",
-    "karma-jasmine": "^1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "^5.4.1",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "^2.6.2"
+    "ts-node": "~7.0.1",
+    "tslint": "~5.8.0",
+    "typescript": "3.1.6"
   },
   "snyk": true
 }
