@@ -1,10 +1,10 @@
-import {Injectable} from '@angular/core';
-import {HttpClient} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { Observable } from 'rxjs';
 
-import {ApiSubmitter} from '../models/api-submitter.model';
-import {ApiService} from './api-service';
-import {ServiceBase} from './service-base';
+import { ApiSubmitter } from '../models/api-submitter.model';
+import { ApiService } from './api-service';
+import { ServiceBase } from './service-base';
 
 /**
  * Service for obtaining submitter objects.
